+++ flatlaf-jide-oss/src/main/java/com/formdev/flatlaf/jideoss/FlatJideOssDefaultsAddon.java
@@ -46,6 +46,11 @@ public class FlatJideOssDefaultsAddon
 		return super.getDefaults( lafClass );
 	}
 
+	@Override
+	public int getPriority() {
+		return 11;
+	}
+
 	//---- class FlatJideUIDefaultsCustomizer ---------------------------------
 
 	/**